addInitScript in qawolf.register since it happens before pages load, so it should not throw any errors with the addInitScript
- Remove the ability to independently call indexPages. Just call qawolf.register and indexPages and ContextEventCollector work. One way to do things
