+++ src/create-code/ContextEventCollector.ts
@@ -1,19 +1,10 @@
 import Debug from 'debug';
 import { EventEmitter } from 'events';
 import { BrowserContext } from 'playwright-core';
-import { webScript } from '../web/addScript';
-import { loadConfig } from '../config';
+import { IndexedPage, isRegistered } from '../utils/context/register';
 import { ElementEvent } from '../types';
-import { indexPages, IndexedPage } from '../utils';
 
 const debug = Debug('qawolf:ContextEventCollector');
-
-const ignoreError = (error: Error) => {
-  // when an iframe closes it will throw an error we can ignore
-  return !!['Session closed', 'Target closed'].find((ignore) =>
-    error.message.includes(ignore),
-  );
-};
 
 export class ContextEventCollector extends EventEmitter {
   readonly _attribute: string;
@@ -23,18 +14,19 @@ export class ContextEventCollector extends EventEmitter {
     context: BrowserContext,
   ): Promise<ContextEventCollector> {
     const collector = new ContextEventCollector(context);
-    await collector._start();
+    await collector._emitEvents();
     return collector;
   }
 
   protected constructor(context: BrowserContext) {
     super();
-    this._attribute = loadConfig().attribute;
     this._context = context;
   }
 
-  async _start(): Promise<void> {
-    await indexPages(this._context);
+  async _emitEvents(): Promise<void> {
+    if (!isRegistered(this._context)) {
+      throw new Error('Use qawolf.register(context) first');
+    }
 
     await this._context.exposeBinding(
       'qawElementEvent',
@@ -45,31 +37,5 @@ export class ContextEventCollector extends EventEmitter {
         this.emit('elementevent', event);
       },
     );
-
-    const script =
-      '(() => {\n' +
-      webScript +
-      `\nnew qawolf.PageEventCollector({ attribute: ${JSON.stringify(
-        this._attribute,
-      )} });\n` +
-      '})();';
-
-    try {
-      await Promise.all(
-        this._context.pages().map((page) => page.evaluate(script)),
-      );
-    } catch (error) {
-      if (!ignoreError(error)) throw error;
-
-      debug('evaluate error we ignored: ' + error.message);
-    }
-
-    try {
-      await this._context.addInitScript(script);
-    } catch (error) {
-      if (!ignoreError(error)) throw error;
-
-      debug('addInitScript error we ignored: ' + error.message);
-    }
   }
 }
